+++ packages/nwayo/src/index.js
@@ -1,11 +1,15 @@
 //--------------------------------------------------------
 //-- ESLint config
 //--------------------------------------------------------
-import path   from 'path';
-import pkgDir from 'pkg-dir';
-import fss    from '@absolunet/fss';
+import deepMerge      from 'deepmerge';
+import prettierConfig from 'eslint-config-prettier';
+import path           from 'path';
+import pkgDir         from 'pkg-dir';
+import fss            from '@absolunet/fss';
 
-module.exports = fss.readYaml(path.join(
+const shareableConfig = fss.readYaml(path.join(
 	pkgDir.sync(__dirname),
 	'shareable-config.yaml'
 ));
+
+module.exports = deepMerge(shareableConfig, prettierConfig);